[UPDATE] 소모임 수정/ 1:1 채팅

+++ src/main/java/com/playground/th/controller/dto/responseDto/ResponseDto.java
@@ -9,6 +9,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ResponseDto {
-    private int result;
+    private Long result;
     private String message;
 }
